backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Hannah Lamarche"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
motor = "^3.5.0"
loguru = "^0.7.2"
pydantic-settings = "^2.3.4"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.2"
pytest = "^8.2.2"
anyio = "^4.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
